refact: ticketOption api 분리
