+++ src/features/ticket/api/ticket.ts
@@ -1,5 +1,5 @@
 import { axiosClient } from "../../../shared/types/api/http-client";
-import { CreateTicketRequest, ReadTicketResponse, TicketOptionAnswerRequest, TicketOptionResponse } from "../model/ticketInformation";
+import { CreateTicketRequest, ReadTicketResponse } from "../model/ticketInformation";
 
 export const createTicket = async (data: CreateTicketRequest) => {
   const response = await axiosClient.post('/tickets', data);
@@ -18,12 +18,3 @@ export const deleteTicket = async (ticketId: number) => {
     return response.data;
 }
 
-export const readTicketOptions = async (ticketId: number): Promise<{ isSuccess: boolean; result: TicketOptionResponse[] }> => {
-    const response = await axiosClient.get(`/ticket-options/tickets/${ticketId}`);
-    return response.data;
-};
-
-export const createTicketOptionAnswers = async (answers: TicketOptionAnswerRequest) => {
-  const response = await axiosClient.post("/ticket-option-answers", answers);
-  return response.data;
-};
